+++ ovp_projects/models/apply.py
@@ -4,10 +4,15 @@ from django.utils import timezone
 
 from ovp_projects import emails
 
+apply_status_choices = (
+    ('applied', 'Applied'),
+    ('unapplied', 'Canceled'),
+)
+
 class Apply(models.Model):
   user = models.ForeignKey('ovp_users.User', blank=True, null=True, verbose_name=_('user'))
   project = models.ForeignKey('ovp_projects.Project', verbose_name=_('project'))
-  status = models.CharField(_('status'), max_length=30)
+  status = models.CharField(_('status'), max_length=30, choices=apply_status_choices, default="applied")
   date = models.DateTimeField(_('created date'), auto_now_add=True, blank=True)
   canceled = models.BooleanField(_("canceled"), default=False)
   canceled_date = models.DateTimeField(_("canceled date"), blank=True, null=True)